requirements.txt
requests==2.28.2
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0